Split create-flow test into two
